Current `cmd/main.go` file is messy
Extract `account` command into a separate file.
