+++ pkg/vanity/selector_combi.go
@@ -31,7 +31,7 @@ func (sel seqSelector) Pass(addr keypair.Address) bool {
 func (sel seqSelector) rules() []searchRule {
 	res := []searchRule{{}}
 	for _, it := range sel.items {
-		n := make([]searchRule, 0)
+		var n []searchRule
 		for _, r1 := range res {
 			for _, r2 := range it.rules() {
 				n = append(n, r1.merge(r2))
@@ -65,7 +65,7 @@ func (sel parSelector) rules() []searchRule {
 	if len(sel.items) == 0 {
 		return []searchRule{{}}
 	}
-	res := make([]searchRule, 0)
+	var res []searchRule
 	for _, it := range sel.items {
 	OUTER:
 		for _, r := range it.rules() {
